+++ package.json
@@ -44,6 +44,7 @@
     "eslint-plugin-react-refresh": "^0.4.12",
     "globals": "^15.11.0",
     "prettier": "^3.3.3",
+    "react-error-boundary": "^4.1.2",
     "typescript": "^5.7.2",
     "typescript-eslint": "^8.11.0",
     "vite": "^5.4.11",
